Change LAI value to float instead of byte.

+++ src/OutputAggregatedTable.cs
@@ -28,7 +28,7 @@ namespace Landis.Extension.Output.PnET
                 ISiteVar<float> CohortWoodySenescence = PlugIn.cohorts.GetIsiteVar(x => x.WoodySenescenceSum);
                 ISiteVar<float> FoliageWoodySenescence = PlugIn.cohorts.GetIsiteVar(x => x.FoliageSenescenceSum);
                 ISiteVar<int> CohortAge = PlugIn.cohorts.GetIsiteVar(x => (x.CohortCount >0) ? x.AverageAge : -1);
-                ISiteVar<byte> CohortLAI = PlugIn.cohorts.GetIsiteVar(x => x.CanopyLAImax);
+                ISiteVar<float> CohortLAI = PlugIn.cohorts.GetIsiteVar(x => x.CanopyLAImax);
                 ISiteVar<float> WaterPerSite = PlugIn.cohorts.GetIsiteVar(x => x.WaterMax);
                 ISiteVar<float> SubCanopyRAD = PlugIn.cohorts.GetIsiteVar(x => x.SubCanopyParMAX);
                 ISiteVar<double> Litter = PlugIn.cohorts.GetIsiteVar(x => x.Litter);
